requirements.txt
aiogram~=2.14.1
python-dotenv~=0.18.0
psycopg2-binary==2.9.1
aioredis
loguru
git+https://github.com/KrimsN/pyrogram.git